+++ backend/api/chat-core/builtInTools.js
@@ -1,8 +1,8 @@
-const {db, dbAll} = require('../database/config/database.js');
-const { getPageByUrlAndWebsiteId } = require('../database/queries');
-const {getWebsiteById} = require('../database/queries');
-const { searchFileContent, readFileContent, getFilesByWebsiteId } = require('../database/queries');
-const { logToolCall } = require('../database/logging/toolCalls.js');
+const { dbAll } = require('../../database/config/database.js');
+const { getPageByUrlAndWebsiteId } = require('../../database/queries');
+const { getWebsiteById } = require('../../database/queries');
+const { searchFileContent, readFileContent, getFilesByWebsiteId } = require('../../database/queries');
+const { logToolCall } = require('../../database/logging/toolCalls.js');
 const wsManager = require('./wsManager.js');
 // a set of tools the chatbot can use to find information for the user
 tools = [
