+++ backend/api/chat-core/chatbotResponse.js
@@ -6,23 +6,27 @@ const dotenv = require("dotenv");
 dotenv.config();
 
 const { appendMessageToSession, getSession } = require('./sessions.js');
-const { dbGet } = require('../database/config/database.js');
+const { dbGet } = require('../../database/config/database.js');
 
-const { getPlanFromChatbotId, subtractFromPlan } = require('../database/queries/billing/plans.js');
+const { getPlanFromChatbotId, subtractFromPlan } = require('../../database/queries/billing/plans.js');
 
-const { getSystemPrompt } = require('../database/queries');
-const { getWebsiteByChatbotId } = require('../database/queries');
+const { getSystemPrompt } = require('../../database/queries');
+const { getWebsiteByChatbotId } = require('../../database/queries');
 const { 
     checkAndRenewCredits,
     getMonthlyCredits,
     checkAndSetWarningFlag
-} = require('../database/queries');
+} = require('../../database/queries');
 
-const { logger } = require('../../utils/fileLogger.js');
+const { logger } = require('../utils/fileLogger.js');
 
-const { sendCreditsExhaustedEmail, sendCreditsLowWarningEmail, sendCreditsHalfWarningEmail } = require('../../utils/emailService.js');
+const { 
+    sendCreditsExhaustedEmail, 
+    sendCreditsLowWarningEmail, 
+    sendCreditsHalfWarningEmail 
+} = require('../utils/emailService.js');
 
-const { getUserByPlanId } = require('../database/queries/auth/users.js');
+const { getUserByPlanId } = require('../../database/queries/auth/users.js');
 
 // use tool requires tool name and params
 const { getTools, useTool } = require('./builtInTools.js');
