+++ backend/api/chat-core/utils/rateLimiter.js
@@ -1,4 +1,4 @@
-const { dbGet, dbRun } = require('../../database/config/database.js');
+const { dbGet, dbRun } = require('../../../database/config/database.js');
 
 // Rate limit window in milliseconds
 const WINDOW_MS = 24 * 60 * 60 * 1000; // 24 hours (1 day)
