+++ backend/api/files/textExtractor.js
